adding dictionary package list

+++ mobile/lib/ui/dictionary/dictionary_screen/widgets/word_list_tile.dart
@@ -62,6 +62,7 @@ class WordListTile extends ConsumerWidget {
           ),
           onPressed: () async {
             final db = ref.read(dictionaryDatabaseProvider);
+            if (db == null) return;
             if (isFav) {
               await db.removeFavorite(word.dictId);
             } else {
